+++ core/ledger-client/src/index.ts
@@ -1,10 +1,9 @@
 // Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './ledger-client.js'
 export * from './topology-write-service.js'
 export * from './token-standard-service.js'
-export * as versioned_client from './versioned-ledger-client.js'
+export * from './ledger-client.js'
 export * from './generated-clients/openapi-3.3.0-SNAPSHOT.js'
 export * from './_proto/com/digitalasset/canton/crypto/v30/crypto.js'
 export * from './_proto/com/digitalasset/canton/protocol/v30/topology.js'
