+++ core/ledger-client/src/topology-write-service.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { LedgerClient } from './versioned-ledger-client.js'
+import { LedgerClient } from './ledger-client.js'
 import {
     CryptoKeyFormat,
     SigningKeyScheme,
