--- core/ledger-client/src/versioned-ledger-client.ts
@@ -1,230 +1,0 @@
-// Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import * as v3_3 from './generated-clients/openapi-3.3.0-SNAPSHOT.js'
-
-import * as v3_4 from './generated-clients/openapi-3.4.0-SNAPSHOT.js'
-import createClient, { Client } from 'openapi-fetch'
-import { Logger } from 'pino'
-import { PartyId } from '@canton-network/core-types'
-
-type SupportedVersions = '3.3' | '3.4'
-
-type paths = v3_3.paths | v3_4.paths
-
-// A conditional type that filters the set of OpenAPI path names to those that actually have a defined POST operation.
-// Any path without a POST is excluded via the `never` branch of the conditional
-type PostEndpoint = {
-    [Pathname in keyof paths]: paths[Pathname] extends {
-        post: unknown
-    }
-        ? Pathname
-        : never
-}[keyof paths]
-
-// Given a pathname (string) that has a POST, this helper type extracts the request body type from the OpenAPI definition.
-export type PostRequest<Path extends PostEndpoint> = paths[Path] extends {
-    post: { requestBody: { content: { 'application/json': infer Req } } }
-}
-    ? Req
-    : never
-
-// Given a pathname (string) that has a POST, this helper type extracts the 200 response type from the OpenAPI definition.
-export type PostResponse<Path extends PostEndpoint> = paths[Path] extends {
-    post: { responses: { 200: { content: { 'application/json': infer Res } } } }
-}
-    ? Res
-    : never
-
-// Similar as above, for GETs
-type GetEndpoint = {
-    [Pathname in keyof paths]: paths[Pathname] extends {
-        get: unknown
-    }
-        ? Pathname
-        : never
-}[keyof paths]
-
-// Similar as above, for GETs
-export type GetResponse<Path extends GetEndpoint> = paths[Path] extends {
-    get: { responses: { 200: { content: { 'application/json': infer Res } } } }
-}
-    ? Res
-    : never
-
-export class LedgerClient {
-    // privately manage the active connected version and associated client codegen
-    private readonly clients: Record<SupportedVersions, Client<paths>>
-    private clientVersion: SupportedVersions = '3.3' // default to 3.3 if not provided
-    private currentClient: Client<paths>
-    private initialized: boolean = false
-    private readonly logger: Logger
-
-    // ...
-
-    constructor(
-        baseUrl: URL,
-        token: string,
-        _logger: Logger,
-        version?: SupportedVersions
-    ) {
-        this.logger = _logger.child({ component: 'LedgerClient' })
-        this.clients = {
-            '3.3': createClient<v3_3.paths>({
-                baseUrl: baseUrl.href,
-                fetch: async (url: RequestInfo, options: RequestInit = {}) => {
-                    return fetch(url, {
-                        ...options,
-                        headers: {
-                            ...(options.headers || {}),
-                            Authorization: `Bearer ${token}`,
-                        },
-                    })
-                },
-            }),
-            '3.4': createClient<v3_4.paths>({
-                baseUrl: baseUrl.href,
-                fetch: async (url: RequestInfo, options: RequestInit = {}) => {
-                    return fetch(url, {
-                        ...options,
-                        headers: {
-                            ...(options.headers || {}),
-                            Authorization: `Bearer ${token}`,
-                        },
-                    })
-                },
-            }),
-        }
-
-        this.clientVersion = version ?? this.clientVersion
-        this.currentClient = this.clients[this.clientVersion]
-    }
-
-    public async init() {
-        if (!this.initialized) {
-            // call the /v2/version endpoint on the current client
-
-            const versionFromClient =
-                await this.currentClient.GET('/v2/version')
-
-            // check if the result matches the local version
-            //   if not, but the version is supported, update `this.clientVersion`
-            //   if not, and version is unsupported, throw an error
-
-            //todo fix this
-
-            if (versionFromClient.data?.version.includes('3.3')) {
-                this.clientVersion = '3.3'
-                this.initialized = true
-            } else if (versionFromClient.data?.version.includes('3.4')) {
-                this.clientVersion = '3.4'
-                this.initialized = true
-            } else {
-                throw new Error(
-                    `Unsupported version - found ${versionFromClient.data?.version}`
-                )
-            }
-        }
-    }
-
-    /**
-     * Grants a user the right to act as a party, while ensuring the party exists.
-     *
-     * @param userId The ID of the user to grant rights to.
-     * @param partyId The ID of the party to grant rights for.
-     * @returns A promise that resolves when the rights have been granted.
-     */
-    public async grantUserRights(userId: string, partyId: PartyId) {
-        // Wait for party to appear on participant
-        let partyFound = false
-        let tries = 0
-        const maxTries = 20
-
-        while (!partyFound && tries < maxTries) {
-            const parties = await this.get('/v2/parties')
-            partyFound =
-                parties.partyDetails?.some(
-                    (party) => party.party === partyId
-                ) || false
-
-            await new Promise((resolve) => setTimeout(resolve, 1000))
-            tries++
-        }
-
-        if (tries >= maxTries) {
-            throw new Error('timed out waiting for new party to appear')
-        }
-
-        // Assign user rights to party
-        const result = await this.post(
-            '/v2/users/{user-id}/rights',
-            {
-                identityProviderId: '',
-                userId,
-                rights: [
-                    {
-                        kind: {
-                            CanActAs: {
-                                value: {
-                                    party: partyId,
-                                },
-                            },
-                        },
-                    },
-                ],
-            },
-            {
-                path: {
-                    'user-id': userId,
-                },
-            }
-        )
-
-        if (!result.newlyGrantedRights) {
-            throw new Error('Failed to grant user rights')
-        }
-
-        return
-    }
-
-    public async post<Path extends PostEndpoint>(
-        path: Path,
-        body: PostRequest<Path>,
-        params?: {
-            path?: Record<string, string>
-            query?: Record<string, string>
-        }
-    ): Promise<PostResponse<Path>> {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any -- (cant align this with openapi-fetch generics :shrug:)
-        const options = { body, params } as any
-
-        const resp = await this.currentClient.POST(path, options)
-        return this.valueOrError(resp)
-    }
-
-    public async get<Path extends GetEndpoint>(
-        path: Path,
-        params?: {
-            path?: Record<string, string>
-            query?: Record<string, string>
-        }
-    ): Promise<GetResponse<Path>> {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any -- (cant align this with openapi-fetch generics :shrug:)
-        const options = { params } as any
-        const resp = await this.currentClient.GET(path, options)
-        return this.valueOrError(resp)
-    }
-
-    private async valueOrError<T>(response: {
-        data?: T
-        error?: unknown
-    }): Promise<T> {
-        if (response.data === undefined) {
-            return Promise.reject(response.error)
-        } else {
-            return Promise.resolve(response.data)
-        }
-    }
-
-    // ... for each method, ensure that the LedgerClient.initialize is true, and throw if not
-}
