+++ sdk/wallet-sdk/src/ledgerController.ts
@@ -9,7 +9,6 @@ import {
     Types,
     awaitCompletion,
     promiseWithTimeout,
-    versioned_client,
 } from '@canton-network/core-ledger-client'
 import {
     signTransactionHash,
@@ -40,7 +39,6 @@ export type WrappedCommand<
  */
 export class LedgerController {
     private readonly client: LedgerClient
-    private readonly versionedLedgerClient: versioned_client.LedgerClient
     private readonly userId: string
     private partyId: PartyId | undefined
     private synchronizerId: PartyId | undefined
@@ -54,11 +52,7 @@ export class LedgerController {
      */
     constructor(userId: string, baseUrl: URL, token: string) {
         this.client = new LedgerClient(baseUrl, token, this.logger)
-        this.versionedLedgerClient = new versioned_client.LedgerClient(
-            baseUrl,
-            token,
-            this.logger
-        )
+        this.client.init()
         this.userId = userId
         return this
     }
